libage_crypto

Function encrypt_multiple

pub fn encrypt_multiple(
    recipients: &[Recipient],
    secret: &Secret,
) -> Result<EncryptedPayload>

Encrypts secret for multiple recipients at once.

The resulting EncryptedPayload can be decrypted with any of the corresponding identities (private keys).

Examples

let (r1, id1) = generate_keypair()?;
let (r2, _)   = generate_keypair()?;
let secret = Secret::new(b"multi-recipient".to_vec());

let encrypted = encrypt_multiple(&[r1, r2], &secret)?;
let decrypted = AgeCrypto.decrypt(&id1, &encrypted)?;
assert_eq!(decrypted.as_bytes(), secret.as_bytes());

Errors

Returns Err if any of the provided recipients is invalid, encryption fails, or librage returns an unexpected response.